refactor(matchers): split value_matcher into json_matcher and primitive_matcher
