+++ src/matchers/elements_are_matcher.rs
@@ -65,7 +65,7 @@ macro_rules! __json_elements_are {
 
 #[doc(hidden)]
 pub mod internal {
-    use crate::matchers::value_matcher::internal::JsonMatcher;
+    use crate::matchers::json_matcher::internal::JsonMatcher;
     use googletest::description::Description;
     use googletest::matcher::{Matcher, MatcherBase, MatcherResult};
     use serde_json::Value;
